fix build: resolve duplicate symbols, type mismatches, add file utils
- Remove libmsi_enums/config_h from msi-interop target to avoid
  duplicate symbol definitions with extracted libmsi objects
- Change copy_string_to_bufA/W pcch param from unsigned int* to
  unsigned long* for Windows LPDWORD compatibility
- Implement MsiGetFileHashA/W using GLib MD5 checksum
- Implement MsiGetFileVersionA/W with PE version info parsing
- Remove corresponding stubs from stubs.c

+++ msi-interop/string_utils.c
@@ -52,7 +52,7 @@ utf8_to_utf16(const char *utf8, size_t *out_len)
 }
 
 unsigned int
-copy_string_to_bufA(const char *str, char *buf, unsigned int *pcch)
+copy_string_to_bufA(const char *str, char *buf, unsigned long *pcch)
 {
     if (str == NULL)
         str = "";
@@ -103,7 +103,7 @@ copy_string_to_bufA(const char *str, char *buf, unsigned int *pcch)
 }
 
 unsigned int
-copy_string_to_bufW(const char *str, WCHAR *buf, unsigned int *pcch)
+copy_string_to_bufW(const char *str, WCHAR *buf, unsigned long *pcch)
 {
     if (str == NULL)
         str = "";
